chore: extract hooks
