+++ strapi/src/admin/PreviewProxy.tsx
@@ -8,53 +8,8 @@ import {
   SingleSelect,
   SingleSelectOption,
 } from "@strapi/design-system";
-import { unstable_useDocument as useDocument } from "@strapi/strapi/admin";
-
-const ExpoPreview = () => {
-  const qrCodeSrc = React.useMemo(() => {
-    const qrCodeUrl = new URL("https://qr.expo.dev/eas-update");
-    qrCodeUrl.searchParams.append(
-      "projectId",
-      "4327bdd6-9794-49d7-9b95-6a5198afd339",
-    );
-    qrCodeUrl.searchParams.append("runtimeVersion", "1.0.0");
-    qrCodeUrl.searchParams.append("channel", "default");
-    return qrCodeUrl.toString();
-  }, []);
-
-  return (
-    <Flex
-      display="flex"
-      alignItems="center"
-      justifyContent="center"
-      height="100%"
-      width="100%"
-    >
-      <img src={qrCodeSrc} alt="Expo QR Code" />
-    </Flex>
-  );
-};
-
-const devices = [
-  {
-    name: "Web (full screen)",
-    width: "100%",
-    height: "100%",
-  },
-  {
-    name: "Web (iPhone 5)",
-    width: "320px",
-    height: "568px",
-  },
-  {
-    name: "Web (iPad)",
-    width: "768px",
-    height: "1024px",
-  },
-  {
-    name: "Native mobile app",
-  },
-];
+import { ExpoPreview } from "./utils/ExpoPreview";
+import { useDeviceSelector, useUpdateHighlighter } from "./utils/hooks";
 
 const PreviewProxy = () => {
   const [searchParams] = useSearchParams();
@@ -62,51 +17,10 @@ const PreviewProxy = () => {
     Object.fromEntries(searchParams);
   const previewURL = `${clientUrl}/api/preview?${new URLSearchParams(params).toString()}`;
 
-  const previousDocument = React.useRef<any>(undefined);
-  const iframe = React.useRef<HTMLIFrameElement>(null);
+  const { devices, selectedDevice, handleDeviceChange, isMobileApp } =
+    useDeviceSelector();
 
-  const { refetch } = useDocument({
-    collectionType:
-      kind === "collectionType" ? "collection-types" : "single-types",
-    model: params.model,
-    documentId,
-    params: { locale: params.locale },
-  });
-
-  React.useEffect(() => {
-    const handleMessage = async (message) => {
-      if (message.data.type === "strapiUpdate") {
-        const response = await refetch();
-        const document = response.data.data;
-
-        let changedFields: Array<string> = [];
-        if (document != null && previousDocument.current !== document) {
-          // Get the diff of the previous and current document, find the path of changed fields
-          changedFields = Object.keys(document).filter(
-            (key) => document[key] !== previousDocument.current?.[key],
-          );
-        }
-
-        iframe.current?.contentWindow?.postMessage(
-          { ...message.data, changedFields },
-          new URL(iframe.current.src).origin,
-        );
-
-        previousDocument.current = document;
-      }
-    };
-
-    // Add the event listener
-    window.addEventListener("message", handleMessage);
-
-    // Cleanup the event listener on unmount
-    return () => {
-      window.removeEventListener("message", handleMessage);
-    };
-  }, [refetch]);
-
-  const [selectedDevice, setSelectedDevice] = React.useState(0);
-  const device = devices[selectedDevice];
+  const { iframe } = useUpdateHighlighter();
 
   return (
     <Portal>
@@ -121,27 +35,22 @@ const PreviewProxy = () => {
       >
         <Flex gap={4} justifyContent="center" padding={2}>
           <Typography>Preview on:</Typography>
-          <SingleSelect
-            aria-label="Select device"
-            placeholder="Select device"
-            value={selectedDevice}
-            onChange={(e) => setSelectedDevice(Number(e))}
-          >
-            {devices.map((device, index) => (
-              <SingleSelectOption key={device.name} value={index}>
+          <SingleSelect value={selectedDevice.id} onChange={handleDeviceChange}>
+            {devices.map((device) => (
+              <SingleSelectOption key={device.name} value={device.id}>
                 {device.name}
               </SingleSelectOption>
             ))}
           </SingleSelect>
         </Flex>
-        {selectedDevice === devices.length - 1 ? (
+        {isMobileApp ? (
           <ExpoPreview />
         ) : (
           <Box
             tag="iframe"
             src={previewURL}
-            width={device.width}
-            height={device.height}
+            width={selectedDevice.width}
+            height={selectedDevice.height}
             marginLeft="auto"
             marginRight="auto"
             display="block"
